perf: :zap: shorten files and removed duplicated code
